+++ tpl/internal/go_templates/htmltemplate/exec_test.go
@@ -1724,7 +1724,7 @@ var v = "v";
 `
 
 func TestEscapeRace(t *testing.T) {
-	t.Skip("this test currently fails with -race; see issue #39807")
+	// t.Skip("this test currently fails with -race; see issue #39807")
 
 	tmpl := New("")
 	_, err := tmpl.New("templ.html").Parse(raceText)
